update testing/validation pipelines and forge fmt all files (#13)
* forge fmt

* update workflow to fuzz 10,000 times

* update slither workflow version

* use --fuzz-runs everywhere instead of environment variable

+++ src/test/Xorkle.t.sol
@@ -7,6 +7,7 @@ import "forge-std/Test.sol";
 contract ContractTest is DSTest {
     Xorkle m;
     Vm vm = Vm(HEVM_ADDRESS);
+
     function setUp() public {
         m = new Xorkle();
     }
@@ -16,7 +17,7 @@ contract ContractTest is DSTest {
         bytes32 hNaive = keccak256(abi.encode(left ^ right));
         assertEq(hAssem, hNaive);
     }
-    
+
     function testGenerateProof(bytes32[] memory data, uint256 node) public {
         vm.assume(data.length > 1);
         vm.assume(node < data.length);
@@ -25,7 +26,7 @@ contract ContractTest is DSTest {
         bytes32 valueToProve = data[node];
 
         bytes32 rollingHash = valueToProve;
-        for(uint i = 0; i < proof.length; ++i){
+        for (uint256 i = 0; i < proof.length; ++i) {
             rollingHash = m.hashLeafPairs(rollingHash, proof[i]);
         }
         assertEq(rollingHash, root);
@@ -51,7 +52,7 @@ contract ContractTest is DSTest {
 
     function testWontGetRootSingleLeaf() public {
         bytes32[] memory data = new bytes32[](1);
-        data[0] = bytes32(0x0); 
+        data[0] = bytes32(0x0);
         vm.expectRevert("won't generate root for single leaf");
         m.getRoot(data);
     }
@@ -64,7 +65,7 @@ contract ContractTest is DSTest {
     }
 
     function valueNotInArray(bytes32[] memory data, bytes32 value) public pure returns (bool) {
-        for (uint i = 0; i < data.length; ++i) {
+        for (uint256 i = 0; i < data.length; ++i) {
             if (data[i] == value) return false;
         }
         return true;
